Move testAddDescription test from OrganizationFolder2Test class to OrganizationFolderTest class and Remove OrganizationFolder2Test class
